requirements.txt
pytest==8.4.1
selenium==4.24.0
allure-pytest==2.15.0
